+++ src/components/Button/index.ts
@@ -11,3 +11,4 @@ export type {
 } from './Button';
 export {Button} from './Button';
 export {ButtonIcon} from './ButtonIcon';
+export {BUTTON_ICON_SIZE_MAP} from './constants';