+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## [7.48.1](https://github.com/gravity-ui/uikit/compare/v7.48.0...v7.48.1) (2026-08-21)
+
+
+### Bug Fixes
+
+* **Drawer:** support outside clicks without veil ([#2787](https://github.com/gravity-ui/uikit/issues/2787)) ([2e01da5](https://github.com/gravity-ui/uikit/commit/2e01da592246c1a6d2d274d68adb1631cbb6459e))
+* **Modal:** constrain scrollable content to overlay ([#2789](https://github.com/gravity-ui/uikit/issues/2789)) ([ff06988](https://github.com/gravity-ui/uikit/commit/ff069886ec73490999d8e124b414cf61aecbe8da))
+
+## [7.48.0](https://github.com/gravity-ui/uikit/compare/v7.47.2...v7.48.0) (2026-08-11)
+
+
+### Features
+
+* **Drawer:** add `disableModal` property ([#2775](https://github.com/gravity-ui/uikit/issues/2775)) ([6963bc7](https://github.com/gravity-ui/uikit/commit/6963bc7085d8f5721eb13056b714d6f5cb2c4e40))
+
+
+### Bug Fixes
+
+* **SegmentedRadioGroup:** changed option-text display type ([#2765](https://github.com/gravity-ui/uikit/issues/2765)) ([a6bb0f4](https://github.com/gravity-ui/uikit/commit/a6bb0f429db390d507c7744840a3901b4b8903b6))
+* **ColorPicker:** fix typings for onOpenChange ([#2772](https://github.com/gravity-ui/uikit/issues/2772)) ([a788085](https://github.com/gravity-ui/uikit/commit/a788085e4d8feb7cfa2bfc129405f8efa3083f91))
+
+## [7.47.2](https://github.com/gravity-ui/uikit/compare/v7.47.1...v7.47.2) (2026-08-02)
+
+
+### Bug Fixes
+
+* **Button:** export icon size map ([#2766](https://github.com/gravity-ui/uikit/issues/2766)) ([9d4fc42](https://github.com/gravity-ui/uikit/commit/9d4fc4287cf1f63da1bd90aa07e8dd289487c108))
+* **useFocusWithin:** call a current version of the blur handler ([#2768](https://github.com/gravity-ui/uikit/issues/2768)) ([bd0f62c](https://github.com/gravity-ui/uikit/commit/bd0f62cc3039c6fc4a3cff8200a351d815dea301))
+
 ## [7.47.1](https://github.com/gravity-ui/uikit/compare/v7.47.0...v7.47.1) (2026-07-24)